+++ CarDatabase/Domain/Models/CarUpdateModel.cs
@@ -4,8 +4,8 @@ namespace CarDatabase.Domain.Models
 {
     public class CarUpdateModel : ICarId, IOwnerId
     {
-        public ulong OwnerId { get; set; }
-        public ulong CarId { get; set; }
+        public int OwnerId { get; set; }
+        public int CarId { get; set; }
         
         public string Brand { get; set; }
         public string Model { get; set; }
